+++ docs/CHANGELOG.md
@@ -8,3 +8,12 @@
 -> Clean project and organized for more readable code.<br />
 -> Removed local string extension class.<br />
 -> Added missing comments.
+
+### **v.1.0.3.1836** 
+-> Was added a new message result `Exception`.<br />
+-> Was added a new methods for 'Exception': `AddException`, `HasAnyExceptions`.<br />
+-> Was added a new method for 'Error result': `HasAnyErrorsOrExceptions`.<br />
+-> Was adjusted the method for 'Error result': `GetFirstError` return non null value.<br />
+-> Was added a new `operators` for: catch `Exception`, bool or T response on return from methods.<br />
+-> Was adjusted the methods from `ResultOfT` for `JoinResults` and added new method `JoinErrorResults`.<br />
+-> Was update libs, cleaned code, optimized and reorganized.<br />